Development
implemented Raise function, create a new rentViewModel instance with the values of the selected one

further worked on print output for billings and rooms

have begun with removal of billing window and moved billingPeriodViewModel to a new View, accessable via the top togglebutton menu

i am gonna change the ui to have only one window in future commits.

maybe i am gonna change the way billings are managed as well. if so,
they probably will be separated from the RentViewModel and stored into a list within FlatViewModel like before.
Unsure about that, because right now i see only one case where this is preferrable and that is the new "add raise" function, which right now works with a referenced BilliingViewModel, the other values are cloned rather than referenced.