development
finished instruction popup and tooltip implementation for now, print will follow later, every button now has a tooltip, the question mark button has a tooltip and a popup, they explain how to use the program and what to do and expect in few sentences. will optimize this later.

minor ui improvements and bug fixes

implemented state persistancy for annual billing and print

further developed displaysummary option in print, it is now working on rents and annual billing, reducing the output to the sums